Add generic login component
