+++ src/lib/components/Login/BasicAuthLogin.js
@@ -1,16 +1,5 @@
 import React, { Component } from 'react';
-import { connect } from 'react-redux';
 import Cookie from 'js-cookie';
-import './Login.scss';
-
-const mapStateToProps = (state, ownProps) => {
-    const { APP } = state;
-    return {
-        appPassword: APP.password,
-        appIcon: APP.appIcon,
-        loginIcon: APP.appLoginIcon,
-    }
-}
 
 export const isLoggedIn = function () {
     return Cookie.get('dsauth') === "true";
@@ -34,16 +23,14 @@ class BasicAuthLogin extends Component {
     }
 
     render() {
-        const { loginIcon, appIcon } = this.props;
-        const loginImage = {
-            background: `url(${loginIcon || appIcon}) no-repeat center center`,
-        };
+        const { loginImageStyle } = this.props;
+
         return (
             <div className="login">
                 <div>
                     <form className="login-form" onSubmit={(e) => { e.preventDefault(); this.handleLogin(this.password.value); }}>
                         <div className="form-group">
-                            <div className="brand-login" style={loginImage}></div>
+                            <div className="brand-login" style={loginImageStyle}></div>
                             {this.state.loginError === true ? <div className="alert alert-danger">Incorrect password.</div> : null}
                             <label htmlFor="password">Enter your Password</label>
                             <input className="form-control" type="password" ref={(input) => { this.password = input; }} autoFocus />
@@ -55,4 +42,4 @@ class BasicAuthLogin extends Component {
     }
 }
 
-export default connect(mapStateToProps)(BasicAuthLogin);
+export default BasicAuthLogin
